Sécurisations des routes

+++ backend/src/routes/comments.ts
@@ -3,10 +3,31 @@ const router = Router();
 import { pool } from "../db/db.js"; 
 import { requireAuth } from "../middleware/authMiddleware.js";
 
-router.get("/", async (req:any, res:any) => {
+router.get("/", requireAuth, async (req:any, res:any) => {;
   try {
+    const { issue_id } = req.query;
+    if (!issue_id) {
+      return res.status(400).json({ error: "issue_id is required" });
+    }
+    const issueResult = await pool.query(
+      "SELECT project_id FROM issues WHERE issue_id = $1",
+      [issue_id]
+    );
+    if (issueResult.rows.length === 0) {
+      return res.status(404).json({ error: "Issue not found" });
+    }
+
+    const membership = await pool.query(
+      "SELECT * FROM project_members WHERE project_id = $1 AND user_id = $2",
+      [issueResult.rows[0].project_id, req.user.id]
+    );
+    if (!membership.rows[0]) {
+      return res.status(403).json({ error: "Forbidden: You must be a project member" });
+    }
+
     const result = await pool.query(
-      "SELECT * FROM comments",
+      "SELECT * FROM comments WHERE issue_id = $1",
+      [issue_id]
     );
     res.json(result.rows);
   } catch (error) {
@@ -40,9 +61,16 @@ router.post("/",requireAuth, async (req:any, res:any) => {
   }
 });
 
-router.delete("/:issue_id/:comment_id", async (req:any, res:any) => {
+router.delete("/:issue_id/:comment_id",requireAuth, async (req:any, res:any) => {
+  const { issue_id, comment_id } = req.params;
   try {
-    const { issue_id, comment_id } = req.params;
+    const pmResult = await pool.query(
+      "SELECT * FROM project_members WHERE project_id = (SELECT project_id FROM issues WHERE issue_id = $1) AND user_id = $2",
+      [issue_id, req.user.id]
+    );
+    if (!pmResult.rows[0]) {
+      return res.status(403).json({ error: "Forbidden: You must be a project member to delete a comment" });
+    }
     const result = await pool.query(
       "DELETE FROM comments WHERE issue_id = $1 AND comment_id = $2 RETURNING *",
       [issue_id, comment_id]
@@ -58,7 +86,7 @@ router.delete("/:issue_id/:comment_id", async (req:any, res:any) => {
   }
 });
 
-router.put("/:issue_id/:comment_id",async (req:any, res:any) => {
+router.put("/:issue_id/:comment_id",requireAuth,async (req:any, res:any) => {
   try {
     const { content } = req.body;
     const { issue_id, comment_id } = req.params;
